1_Data_Pipelines/requirement.txt
pandas==1.5.3
apache-airflow==2.5.3